fix: use local config when env is set to testing/dev
